requirements.txt
pyyaml
requests
BeautifulSoup4
python-dateutil
python-slugify
